+++ bold-pagos-en-linea/include/Components/BoldButtonBlock.php
@@ -76,7 +76,7 @@ class BoldButtonBlock
         $description = $attrs["description"] ? esc_attr($attrs["description"]) : '';
         $bold_color_button = esc_attr($attrs["color"]);
         $bold_size_button = esc_attr($attrs["size"]);
-        $woocommerce_bold_version = "wordpress-block-3.1.9";
+        $woocommerce_bold_version = "wordpress-block-3.2.0";
 
         return BoldCommon::getButtonScript(
             $apiKey,